test: handle twofish errors
